feat():登录、注册、以及相关角色部门的创建绑定
-增加了对角色和部分模块的和登录注册相关的基本功能
-新增DepManage模块、RoleManage模块，把更新时的非空判断移到dao层

+++ level1/src/main/java/com/yealink/level1/bean/Role.java
@@ -22,7 +22,7 @@ public class Role {
 
     private String definition;
 
-    private BigInteger createTime;
+    private long createTime;
 
-    private BigInteger modifyTime;
+    private long modifyTime;
 }
