+++ level1/src/main/java/com/yealink/level1/bean/StaffDepartmentRelation.java
@@ -24,7 +24,7 @@ public class StaffDepartmentRelation {
 
     private String position;
 
-    private BigInteger createTime;
+    private long createTime;
 
-    private BigInteger modifyTime;
+    private long modifyTime;
 }
