+++ level1/src/main/java/com/yealink/level1/bean/StaffRoleRelation.java
@@ -22,8 +22,8 @@ public class StaffRoleRelation {
 
     private String roleId;
 
-    private BigInteger createTime;
+    private long createTime;
 
-    private BigInteger modifyTime;
+    private long modifyTime;
 }
